Rename 'yama.db' to 'yama.database'
